Move syscall helpers to their own file
