+++ pkg/security/probe/syscalls_bpf.go
@@ -1,0 +1,29 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-2020 Datadog, Inc.
+
+// +build linux_bpf
+
+package probe
+
+import (
+	"strings"
+
+	"github.com/iovisor/gobpf/elf"
+)
+
+// cache of the syscall prefix depending on kernel version
+var syscallPrefix string
+
+func getSyscallFnName(name string) string {
+	if syscallPrefix == "" {
+		syscall, err := elf.GetSyscallFnName("open")
+		if err != nil {
+			panic(err)
+		}
+		syscallPrefix = strings.TrimSuffix(syscall, "open")
+	}
+
+	return syscallPrefix + name
+}
