+++ pkg/security/probe/syscalls_nobpf.go
@@ -1,0 +1,12 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-2020 Datadog, Inc.
+
+// +build !linux_bpf
+
+package probe
+
+func getSyscallFnName(name string) string {
+	return name
+}
